+++ src/analyzer/p/p.h
@@ -1,8 +1,6 @@
 #ifndef UNITNOS_P_H_
 #define UNITNOS_P_H_
 
-#include <stdint.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -15,7 +13,7 @@ extern "C" {
  *******************************************************************************/
 typedef struct unitnos_p unitnos_p;
 unitnos_p *unitnos_p_create(void);
-void unitnos_p_set_m(unitnos_p *p, uint16_t m);
+void unitnos_p_set_m(unitnos_p *p, unsigned int m);
 void unitnos_p_add_new_file(unitnos_p *p, const char *file);
 void unitnos_p_destroy(unitnos_p *p);
